transition.ease() no/multiple arguments
Add ellipse to indicate that there can be an unlimited number of arguments. This matches the notation used in d3.ease() to which the arguments are passed.
Make it clear that all arguments are passed to d3.ease() when a string is specified as value.
Correct issue with no arguments case. It was stating that *ease* (the method) was unspecified when *value* (argument) was intended. Re-word to make it more clear that it is talking about the case of passing no arguments.